#25647 #25432 Switched to the new dark theme. Better thread protection.
Switches the houdini engine to use the new dark theme introduced in 
https://github.com/shotgunsoftware/tk-core/pull/80

Also adds protection so that you cannot launch UIs from a non-UI thread.
This is a particular problem on Houdini where the python shell runs in a
separate thread, meaning that if you try to launch any QT UI from the
python console, it is likely to lead to instability or crashes.