#30708 Adds the ability to create a Shotgun shelf for Toolkit commands
This change adds 2 new settings to the houdini engine:

  enable_sg_menu: Allows for enabling/disabling the Shotgun menu
  enable_sg_shelf: Allows for enabling/disabling a new Shotgun houdini shelf

Both settings are 'true' by default. The houdini menu is static during a
session and cannot be changed which has made context switching impractical to
this point. These new settings allow the option to migrate to a shelf which
displays the same actions as buttons. The shelf can be dynamic within a session
and, therefore, allow context switching.

To enable context switching, allowable entity types must be added to the
tk-multi-workfile sg_engity_types setting.

Closes #7