Isolating panel support to houdini 15.0.252+ for now.
Even though it isn't released yet, this is the first version with bug fixes
that make python panels usable for us.

SESI fixed an issue with saving python panels in the default desktop for mac.

They're currently looking into an issue with houdini 14 where calling
setInterface() on a python panel does nothing. If that is fixed, then we can
make panels work in newer versions of 14 as well.