#27481 Updates tk-houdini engine for Houdini 14
Now that houdini 14+ comes with PySide, we can avoid creating and managing a
custom event loop for displaying PyQt/PySide widgets in Houdini. The engine now
checks that Houdini 14+ is running and, if so, bypasses the custom event loop
and avoids searching the system for a PyQt/Pyside to use.  Additional logging
has been added to identify when the integrated PySide is being used and when it
is not. Closes #10.