#26983 Adds dynamic menu
This commit enables dynamic menu building for houdini 15+. Changing the context
inside houdini will cause the menu to be rebuilt with the commands registered
in that environment.

The menu looks somewhat different from the static menu built in older versions
of houdini. This is a result of the current limitations of the dynamic menu
building. SESI has a couple of logged RFEs to provide additional flexibility
to improve the experience. As it is though, the menu should maintain the
relative order of commands users are used to seeing, in addition to matching
the order of commands in the Shotgun shelf.