#36334 Adds support for Recent Files menu
This is an initial implementation using a QTimer to poll for
changes to the current hip file. I'm not 100% sure this is the
best way, but it seems to work. Houdini does not seem to have a
good way to hook into work file changes via callbacks.

These changes modify the dynamic menu system to insert a single
menu item "Not working in a Shotgun context" when no engine can
be found. When clicked, it pops up a warning dialog with additional
information.

For versions of Houdini that don't support dynamic menus, I don't
think anything can be done to disable or hide the menu. Those users
don't expect a dynamic menu, so it is unlikely they'd be relying upon
the Recent Files menu.

I'm waiting on a response from SESI to see if there may be a better
way to do this. If we proceed with this solution, it will need some
thorough QA.