#36334 Adds support for Recent Files menu
Adds context switching support for the 'Recent Files' menu in Houdini.
This allows users to select from the menu and have the Shotgun menu/shelves
update with the appropriate context. Also handles opening a file outside of
a SG project by clearing the shelf and replacing the contents of the SG menu
with one item that indicates the user is not working in a SG context. For
Houdini version 14 and below, the menu remains unchanged as there is no way
to update it dynamically.