#37157 Safer logic for modifying HOUDINI_PATH during bootstrap
This change modifies the houdini engine bootstrap logic to be a little safer
when modifying the HOUDINI_PATH. It now ensures that the additional paths are
not already in HOUDINI_PATH before prepending/appending. This should resolve
issues whereby multiple & could end up in the HOUDINI_PATH which was causing
crashes in some scenarios.

Closes #17