#37948 Use existing path separator, fallback to os.pathsep
This change removes the engine's assumption that ; is always used as the
HOUDINI_PATH separator. The bootstrap logic will now check to see if
HOUDINI_PATH is defined and what separator it is using. If an existing
separator is in use, the engine will continue with that separator as it
prepends and appends additional paths required for engine bootstrap. If no
separator is found, then OS-specific separator will be used.

Closes #19.