For #40555: Bundles httplib/urllib2 from Python 2.7.5. (#26)
Bundles httplib/urllib2 from Python 2.7.5 source and uses it when needed. The situation where this is required is on CentOS 7.x, where H16 bundles an ssl.py that is incompatible with the default system Python install's httplib.py. We also bundle urllib2.py from the same Python version to maintain compatibility with our bundled httplib.