For #41913: Disabled panel support for H16 on OS X.
This is a temporary measure while possible fixes for panel painting issues are discussed with SESI engineers. Our panel apps function properly on OS X in dialog mode, so that is the route we'll have to go for the time being. Panel support remains unchanged on Windows and Linux operating systems.

Also addresses a small styling issue with QCheckBox widgets in H16.