For SG-2898: Fix for slow Houdini launches on Windows. (#33)
This is tied to our shelf population logic, but it doesn't appear to be anything we're doing wrong. Rather, as best as I can tell, it's WHEN our logic is being called during launch that causes Houdini to pause for some time. The fix is to defer the shelf population by an event loop cycle, which seems to resolve the issue without any ill effects.