SG-11269: Fixes slow startup on 17.0.506 (#41)
Fixes an issue on Houdini 17.0.506 where the UI just sits and hangs for 2 minutes at startup before launching. This seems to be similar to the bugfix in #33. I have extended that bugfix to wait for the UI to appear before attempting to set up the shelves. This seems to resolve the issue.