SG-14405 fix to load otls in non gui session (#48)
Huge thanks to @reformstudios for submitting this PR. The change allows otl files to still be loaded by the engine even if a GUI isn't present, for example when running from Houdini's hython.

That has been merged in and refactored a bit. I've also updated the tests so that they can be run on Windows. The loading of otl files can now be tested with hython on both Mac and Windows, and all appears to be good.

Co-authored-by: Patrick Macdonald <reformstudios@users.noreply.github.com>
Co-authored-by: Jean-François Boismenu <jfboismenu@users.noreply.github.com>